add header styles
