+++ src/scss/mixins.scss
@@ -1,2 +1,9 @@
 @import './mixins/breakpoints';
 @import './mixins/grids';
+
+@mixin container {
+  width: calc(100vw - 40px);
+  max-width: 1200px;
+  margin-left: auto;
+  margin-right: auto;
+}
