+++ src/templates/MainTemplate.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
-import { NavBar } from '../components/organisms'
+import { Header } from '../components/organisms'
 
 export function MainTemplate({ children }) {
   return (
     <>
-      <NavBar />
+      <Header />
       {children}
     </>
   )
